requirements.txt
Flask==3.1.1
flask-cors==6.0.1
python-dotenv==1.1.1

langchain==0.3.27
langchain-community==0.3.29
langchain-openai>=0.2.8

faiss-cpu==1.12.0
tiktoken==0.11.0

requests==2.32.5
tqdm==4.67.1
urllib3==2.5.0

